release: GOGUFW 2.0.3 action picker and scrambler

+++ Core/Src/system_py32f071.c
@@ -68,8 +68,8 @@
 */
 uint32_t SystemCoreClock = HSI_VALUE;
 
-const uint32_t AHBPrescTable[16] = {0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 6, 7, 8, 9};
-const uint32_t APBPrescTable[8] =  {0, 0, 0, 0, 1, 2, 3, 4};
+const uint8_t AHBPrescTable[16] = {0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 6, 7, 8, 9};
+const uint8_t APBPrescTable[8] =  {0, 0, 0, 0, 1, 2, 3, 4};
 const uint32_t HSIFreqTable[8] = {4000000U, 8000000U, 16000000U, 22120000U, 24000000U, 4000000U, 4000000U, 4000000U};
 
 /**
